Added Anime schema to model.js, deleted extra files, added models.test.js
